Move calculateWattHours into helpers
So services/energy doesn't have to import from another service.
